next attempt to structured bindings

+++ ngstd/hashtable.hpp
@@ -1102,7 +1102,7 @@ namespace ngcore {
 namespace std
 {
   // structured binding support
-  template <size_t N, typename T>
+  template <int N, typename T>
   struct tuple_size<ngstd::INT<N,T>> : std::integral_constant<std::size_t, N> {};
   template<size_t N, int M, typename T> struct tuple_element<N,ngstd::INT<M,T>> { using type = T; };
 }
